refactor(timeline): split Message widget and populate layer out of mod.rs
Stage 4 of the room_screen split, verbatim code motion from two sources:
the Message widget, MessageAction, the message layout constants, and the
whole message DSL template family (plus the thread-summary/bot-card
color aliases they use) move from mod.rs + dsl.rs into message.rs; the
populate_* render pass (all message kinds, replies, thread summaries,
send state, MessageDeliveryState) moves into populate.rs.

message registers before dsl in the script_mod orchestrator since the
RoomScreen/Timeline templates reference the message family. Verified by
booting the app: zero template resolution errors. Public surface
preserved via pub use re-exports (MessageAction, Message refs,
populate_preview_of_timeline_item). mod.rs: 7,903 -> 5,628 lines.
